+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0] - 2026-07-22
+
+### Changed
+
+- Added support for new `parameters` field in `FunctionDebugInfo` to reflect changes in debugger annotations
+
+### Changed
+
+- Added support for new `parameters` field in `FunctionDebugInfo` to reflect changes in debugger annotations
 
 ## [0.10.0] - 2026-06-30